small changes to aufg7 and test1 finished

+++ Rekursion/src/Aufg7.java
@@ -4,7 +4,7 @@ public class Aufg7 {
      * gibt bei fib(0)=0, bei fib(1)=1 zurück und löst alle höheren n in fib(n-1)+fib(n-2) auf, bis 1 oder 0 zurückgegeben werden
      *
      * */
-    public static int fibonacci(int n) {
+    public static long fibonacci(int n) {
         if (n == 0) {
             return 0;
         } else if (n == 1) {
@@ -19,7 +19,7 @@ public class Aufg7 {
     public static void main(String[] args) {
         int n=0;
         while (n<=20) {
-            int werte = fibonacci(n);
+            long werte = fibonacci(n);
             System.out.println("fib(" + n + "): " + werte);
             n++;
         }
